login device id check

+++ htdocs/api/customer/login.php
@@ -24,6 +24,7 @@
 	
 	$isExist = check_user_mobile($mobile, $device_id);
 	$isExist1 = check_user_mobile_temp($mobile);
+	$isDeviceExist = check_user_device($device_id);
 
 	
 	if($isExist)
@@ -71,44 +72,51 @@
 		}
 		else
 		{
-			// Save data in Temp table
-			$objectSoc = new SocieteTemp($db);
+			if($isDeviceExist){
+				$status_code = '0';
+				$message = 'Device ID already added to another customer!!';
+	
+				$json = array('status_code' => $status_code, 'message' => $message);
+			}else{
+				// Save data in Temp table
+				$objectSoc = new SocieteTemp($db);
 
-			$otp = rand(111111, 999999);
+				$otp = rand(111111, 999999);
 
-			$sql = "DELETE FROM ".MAIN_DB_PREFIX."societe_temp WHERE phone = '".$db->escape($mobile)."'";
-			$resql = $db->query($sql);
+				$sql = "DELETE FROM ".MAIN_DB_PREFIX."societe_temp WHERE phone = '".$db->escape($mobile)."'";
+				$resql = $db->query($sql);
 
-			$sql = "INSERT INTO ".MAIN_DB_PREFIX."societe_temp SET phone = '".$db->escape($mobile)."', statut = '0'";
-			$resql = $db->query($sql);
-			$last_insert = $db->last_insert_id(MAIN_DB_PREFIX."societe_temp");
+				$sql = "INSERT INTO ".MAIN_DB_PREFIX."societe_temp SET phone = '".$db->escape($mobile)."', statut = '0'";
+				$resql = $db->query($sql);
+				$last_insert = $db->last_insert_id(MAIN_DB_PREFIX."societe_temp");
 
-			$sql = "INSERT INTO ".MAIN_DB_PREFIX."socpeople_temp SET fk_soc = '".(int)$last_insert."', phone = '".$db->escape($mobile)."', phone_mobile = '".$db->escape($mobile)."', otp = '".$db->escape($otp)."', statut = '0'";
-			$resql = $db->query($sql);
-			$last_insert_people = $db->last_insert_id(MAIN_DB_PREFIX."socpeople_temp");
+				$sql = "INSERT INTO ".MAIN_DB_PREFIX."socpeople_temp SET fk_soc = '".(int)$last_insert."', phone = '".$db->escape($mobile)."', phone_mobile = '".$db->escape($mobile)."', otp = '".$db->escape($otp)."', statut = '0'";
+				$resql = $db->query($sql);
+				$last_insert_people = $db->last_insert_id(MAIN_DB_PREFIX."socpeople_temp");
+				
+				/*
+				$object = new ContactTemp($db);
+
+				$object->firstname = $object->lastname = $object->priv = "";
+				$object->statut = 0;
+
+				$object->create($tempUser);*/
+
+				$status_code = '1';
+				$message = 'New account has been created!';
+
+				$smsmessage = str_replace(" ", "%20", "Dear, Your OTP for login is ".$otp.". Please DO NOT share OTP.");
+				$SENDERID = $conf->global->MAIN_MAIL_SMS_FROM;
+				$PHONE = $mobile;
+				$MESSAGE = $smsmessage;
+				$url = "http://opensms.microprixs.com/api/mt/SendSMS?user=jmvd&password=jmvd&senderid=".$SENDERID."&channel=TRANS&DCS=0&flashsms=0&number=".$PHONE."&text=".$MESSAGE."&route=15";
 			
-			/*
-			$object = new ContactTemp($db);
-
-			$object->firstname = $object->lastname = $object->priv = "";
-			$object->statut = 0;
-
-			$object->create($tempUser);*/
-
-			$status_code = '1';
-			$message = 'New account has been created!';
-
-			$smsmessage = str_replace(" ", "%20", "Dear, Your OTP for login is ".$otp.". Please DO NOT share OTP.");
-			$SENDERID = $conf->global->MAIN_MAIL_SMS_FROM;
-			$PHONE = $mobile;
-			$MESSAGE = $smsmessage;
-			$url = "http://opensms.microprixs.com/api/mt/SendSMS?user=jmvd&password=jmvd&senderid=".$SENDERID."&channel=TRANS&DCS=0&flashsms=0&number=".$PHONE."&text=".$MESSAGE."&route=15";
-		
-			require_once DOL_DOCUMENT_ROOT.'/core/class/CSMSSend.class.php';
-			$smsfile = new CSMSSend($url);
-			$result = $smsfile->sendSMS();
-			
-			$json = array('status_code' => $status_code, 'message' => $message, 'user_id' => "", 'user_otp' => "".$otp, 'fullname' => '', 'mobile' => "".$mobile, 'customer_type' => 'new');
+				require_once DOL_DOCUMENT_ROOT.'/core/class/CSMSSend.class.php';
+				$smsfile = new CSMSSend($url);
+				$result = $smsfile->sendSMS();
+				
+				$json = array('status_code' => $status_code, 'message' => $message, 'user_id' => "", 'user_otp' => "".$otp, 'fullname' => '', 'mobile' => "".$mobile, 'customer_type' => 'new');
+			}	
 		}
 	}
